third_party/blink/web_tests/external/wpt/css/css-animations/KeyframeEffect-setKeyframes.tentative-expected.txt

This is a testharness.js-based test.
PASS KeyframeEffect.setKeyframes() causes subsequent changes to @keyframes rules to be ignored
FAIL KeyframeEffect.setKeyframes() causes subsequent changes to animation-timing-function to be ignored assert_equals: Easing should be the easing set by the API expected "linear" but got "ease-in"
FAIL KeyframeEffect.setKeyframes() should NOT cause subsequent changes to @keyframes rules to be ignored if it threw assert_equals: value for 'left' on Keyframes reflect the value set via style should match expected "300px" but got "100px"
Harness: the test ran to completion.
